fiz o login funcionar mas tem que melhorar o logout da ws
