+++ vite-project/src/components/useWebSocket.tsx
@@ -10,7 +10,7 @@ const useWebSocket = (token: string,) => {
   const [data, setData] = useState(null);
   const ws = useRef<WebSocket | null>(null);
   const timer = useRef<NodeJS.Timeout | null>(null);
-  const { isAdmin } = useAccount();
+  const account  = useAccount();
 
   useEffect(() => {
     const currentUrl = window.location.hostname;
@@ -28,14 +28,18 @@ const useWebSocket = (token: string,) => {
       ws.current.onopen = () => {
         if (timer.current) clearTimeout(timer.current);
         console.log("WebSocket connection opened");
-        console.log("isAdmin", isAdmin);
-        ws.current?.send(JSON.stringify({ api: isAdmin ? "admin" : "user", mt: "UserSession" }));
+        console.log(account.isAdmin);
+        console.log(account);
+        ws.current?.send(JSON.stringify({ api: account.isAdmin ? "admin" : "user", mt: "UserSession" }));
     };
 
       ws.current.onclose = (event) => {
         console.log("WebSocket connection closed:", event.code, event.reason);
         if (timer.current) clearTimeout(timer.current as NodeJS.Timeout);
-        timer.current = setTimeout(connect, 10000); // Tenta reconectar após 10 segundos
+        timer.current = setTimeout(() => {
+          console.log("Reconnecting WebSocket...");
+          connect();
+        }, 5000);
       };
 
       ws.current.onerror = (event) => {
@@ -56,8 +60,9 @@ const useWebSocket = (token: string,) => {
     };
   }, [token,]);
   const closeConnection = useCallback(() => {
-    // Feche a conexão WebSocket aqui
-    // ...
+    if (timer.current) clearTimeout(timer.current);
+    if (ws.current) ws.current.close();
+    if(ws.current) ws.current.onclose
   }, []);
 
   return { data, closeConnection };
