+++ vite-project/src/pages/admin/ButtonsAdmin.tsx
@@ -24,6 +24,7 @@ interface User {
 export default function Buttons() {
   const [users, setUsers] = useState<User[]>([]);
   const [selectedUser, setSelectedUser] = useState<User | null>(null);
+ 
 
   useEffect(() => {
     const fetchUsers = async () => {
